Crate rust_monopoly

CIS 198 - Rust Programming Author: Jack Harkins Project Name: Rust-Monopoly

Reexports

extern crate std as std;

extern crate piston;

extern crate graphics;

extern crate glutin_window;

extern crate opengl_graphics;

extern crate rand;

use std::prelude::v1::*;

Modules

board
Board keeps track of most of the game state local to the Monopoly board, including the spaces (and properties contained within them), the players, the card decks, and the background board image.

cards
Cards are used whenever a player lands on a Chance or a Community Chest space. They are meant to be randomly ordered and drawn from a Vector in the Board struct.

game
The Game struct stores the entirety of the game state, including the game board, whether the game is running or not, the current state of a player's turn, and the player's command. It also stores the window in which game objects are drawn.

player
Player stores the player's properties, their token color, the space the player is currently landed on, the last creditor the player paid rent to, whether the player is in jail, and whether the player is currently up for his turn.

property
Contains the majority of the state associated with owning a property, such as its name, the number of houses and hotels held on it, its base rent, and the color group it belongs to.

space
Space stores the player(s) who currently are landed on the space, if any. It also acts as a container for Properties since every property is a space on the board (the converse isn't necessarily true)

Functions

main
The main entry point for the application