Module rust_monopoly::board

Board keeps track of most of the game state local to the Monopoly board, including the spaces (and properties contained within them), the players, the card decks, and the background board image.

Most of the methods in Board do not take in a current player argument since this is already determined by the player_turn field in Board.

Reexports

extern crate rand;

extern crate graphics;

extern crate glutin_window;

extern crate opengl_graphics;

extern crate piston;

extern crate core;

use std::prelude::v1::*;

use std::io::{self, Write};

use std::collections::BTreeMap;

use std::rc::Rc;

use std::cell::RefCell;

use rand::Rng;

use piston::input::*;

use opengl_graphics::{GlGraphics, Texture};

use std::path::Path;

use super::player::*;

use super::property::*;

use super::game::*;

use super::cards::*;

use super::space::*;

Structs

Board

Constants

MAX_NUM_PLAYERS

NUM_SPACES

Functions

confirm_prompt
Get a yes/no answer from stdin

get_dice_roll_12
Simulate a dice roll: return an integer between 2 and 12, inclusive

get_dice_roll_6

get_int
Get an integer from stdin

get_num_players
Get the desired number of players

get_string

get_token_color

get_turn_action