Module rust_monopoly::game

The Game struct stores the entirety of the game state, including the game board, whether the game is running or not, the current state of a player's turn, and the player's command. It also stores the window in which game objects are drawn.

The game proceeds as follows: 1. Player prompted for action. 2. Player enters key corresponding to action in the console. 3. Game executes the player's action. 4. If the game requires a response to the action (such as landing on unowned property), the player will be prompted to respond by typing a key in the window. 5. After the player's turn is finished, the game will check if they are bankrupt. If so, the player's assets will be transferred to the creditor (or the bank if there is no creditor). The bankrupt player will no longer be able to trade properties, collect rents, or otherwise participate in the game. 6. The last remaining player wins the game.

During the course of the game, the Game will notify its Board that the next player is ready to begin his turn via board.start_turn(). When the player lands, the Board will return a LandAction to the Game, representing the type of action taken upon landing (buying/selling/ renting/etc); see LandAction for more details.

Reexports

extern crate opengl_graphics;

extern crate piston;

extern crate glutin_window;

use std::prelude::v1::*;

use std::collections::BTreeMap;

use glutin_window::GlutinWindow;

use std::process;

use piston::window::WindowSettings;

use piston::event_loop::*;

use piston::input::*;

use std::rc::Rc;

use std::cell::RefCell;

use opengl_graphics::{GlGraphics, OpenGL};

use super::board::*;

use super::player::*;

use super::property::*;

use super::space::*;

Structs

Game

Enums

GameState
Master game state: is the game running/set up/over?

TurnCommand
Represents a player's choice of action during their turn

TurnState
Represents the different stages in a player's turn

Constants

GO_SALARY

INCOME_TAX_AMT

LUXURY_TAX_AMT

WINDOW_HEIGHT

WINDOW_WIDTH

Traits

Render
Objects that can be drawn to the screen with the Piston/OpenGL framework