Module rust_monopoly::space

Space stores the player(s) who currently are landed on the space, if any. It also acts as a container for Properties since every property is a space on the board (the converse isn't necessarily true)

Reexports

extern crate rand;

extern crate graphics;

extern crate glutin_window;

extern crate opengl_graphics;

extern crate piston;

extern crate core;

use std::prelude::v1::*;

use std::rc::Rc;

use std::cell::RefCell;

use piston::input::*;

use opengl_graphics::GlGraphics;

use super::player::*;

use super::property::*;

use super::game::*;

Structs

Space
Represents a space on the board that players can land on (not necessarily a property)

Enums

SpaceEnum
Represents the type of a space (property, tax, card draw, jail, etc.)

Constants

ATL_AVE

BALT_AVE

BDWK

BO_RR

CHANCE_BOT

CHANCE_RIGHT

CHANCE_TOP

COMM_CHEST_BOT

COMM_CHEST_LEFT

COMM_CHEST_RIGHT

CONN_AVE

ELEC_UTIL

FREE_PARKING

GO
Space indices

GO_TO_JAIL

IL_AVE

INCOME_TAX

IN_AVE

JAIL

KY_AVE

LUXURY_TAX

MAR_GARD

MED_AVE

NC_AVE

NY_AVE

ORIENT_AVE

PAC_AVE

PARK_PL

PA_AVE

PA_RR

READING_RR

SL_RR

STATES_AVE

ST_CHAR_PL

ST_JAMES_PL

TN_AVE

VA_AVE

VENTNOR_AVE

VERM_AVE

WATER_UTIL